@@ -1,43 +1,65 @@
 {
   "name": "@geeebe/api",
-  "version": "1.0.51",
-  "description": "",
+  "version": "2.0.0",
   "private": false,
+  "description": "Koa api helpers",
+  "repository": {
+    "type": "git",
+    "url": "git@github.com:gregbacchus/geeebe-api.git"
+  },
+  "license": "ISC",
+  "author": "Greg Bacchus",
   "main": "./dist/index.js",
   "types": "./src/index.ts",
   "scripts": {
-    "test": "./node_modules/.bin/jest --coverage",
-    "build": "rm -rf dist && npx tsc --outDir dist --sourceMap",
-    "prepublishOnly": "npm run lint && npm run build && npm version patch",
-    "lint": "./node_modules/.bin/tslint --config tslint.json -t verbose src/**/*.ts",
+    "build": "rm -rf dist && ./node_modules/.bin/tsc --outDir dist --sourceMap",
+    "lint": "./node_modules/.bin/eslint src/ --ext .ts,.tsx",
+    "prepublishOnly": "npm run lint -s && npm run build -s && npm version patch",
     "postpublish": "git push",
-    "start": "node -r ts-node/register src/index.ts",
-    "watch": "npx nodemon"
-  },
-  "author": "Greg Bacchus",
-  "license": "ISC",
-  "devDependencies": {
-    "@types/jest": "^24.0.23",
-    "@types/node": "^12.12.14",
-    "jest": "^24.9.0",
-    "nodemon": "^2.0.1",
-    "supertest": "^4.0.2",
-    "ts-jest": "^24.2.0",
-    "ts-node": "^8.5.4",
-    "tslint": "^5.20.1",
-    "typescript": "^3.7.3"
+    "test": "./node_modules/.bin/jest --verbose --coverage --runInBand --passWithNoTests"
   },
   "dependencies": {
-    "@geeebe/common": "^1.0.75",
-    "@geeebe/logging": "^1.0.23",
-    "@geeebe/service": "^1.0.52",
-    "@types/koa": "^2.11.0",
-    "@types/koa-router": "^7.0.42",
+    "@geeebe/common": "^1.0.77",
+    "@geeebe/logging": "^1.0.26",
+    "@geeebe/service": "^2.0.4",
+    "@koa/router": "^10.0.0",
+    "@types/koa": "^2.13.3",
+    "@types/koa-bodyparser": "^4.3.1",
+    "@types/koa-compress": "^4.0.1",
+    "@types/koa-conditional-get": "^2.0.0",
+    "@types/koa-etag": "^3.0.0",
+    "@types/koa-helmet": "^6.0.3",
+    "@types/koa-static": "^4.0.1",
+    "@types/koa__router": "^8.0.4",
     "@types/lru-cache": "^5.1.0",
-    "axios": "^0.19.0",
-    "jose": "^1.16.0",
-    "koa": "^2.11.0",
-    "koa-router": "^7.4.0",
-    "lru-cache": "^5.1.1"
+    "axios": "^0.21.1",
+    "helmet": "^4.6.0",
+    "jose": "^3.12.3",
+    "koa": "^2.13.1",
+    "koa-bodyparser": "^4.3.0",
+    "koa-compress": "^5.0.1",
+    "koa-etag": "^4.0.0",
+    "koa-helmet": "^6.1.0",
+    "koa-static": "^5.0.0",
+    "lru-cache": "^6.0.0",
+    "prom-client": "^13.1.0",
+    "validata-koa": "^3.0.3"
+  },
+  "devDependencies": {
+    "@types/jest": "^26.0.23",
+    "@types/node": "^15.12.2",
+    "@types/supertest": "^2.0.11",
+    "@typescript-eslint/eslint-plugin": "^4.26.1",
+    "@typescript-eslint/eslint-plugin-tslint": "^4.26.1",
+    "@typescript-eslint/parser": "^4.26.1",
+    "eslint": "^7.28.0",
+    "eslint-plugin-filenames": "^1.3.2",
+    "eslint-plugin-prefer-arrow": "^1.2.3",
+    "jest": "^27.0.4",
+    "nodemon": "^2.0.7",
+    "supertest": "^6.1.3",
+    "ts-jest": "^27.0.3",
+    "ts-node": "^10.0.0",
+    "typescript": "^4.3.2"
   }
 }
