@@ -1,76 +1,46 @@
 import { Statuses, Time } from '@geeebe/common';
 import { logger, Logger } from '@geeebe/logging';
+import * as Router from '@koa/router';
 import axios from 'axios';
-import { ConsumeKeyInput, JSONWebKeySet, JWKS, JWT } from 'jose';
+import { createSecretKey } from 'crypto';
+import { createRemoteJWKSet } from 'jose/jwks/remote';
+import { JWSHeaderParameters, JWTPayload } from 'jose/jwt/unsecured';
+import { GetKeyFunction, jwtVerify, JWTVerifyGetKey, JWTVerifyOptions, KeyLike } from 'jose/jwt/verify';
+import { FlattenedJWSInput } from 'jose/webcrypto/types';
 import { Context, Middleware } from 'koa';
+import { ApiContext, RequestHeaders } from './api';
 
 import Application = require('koa');
-import Router = require('koa-router');
 import LRU = require('lru-cache');
 
 const TOKEN_EXTRACTOR = /^Bearer (.*)$/;
 const debug = logger.child({ module: 'api:authorization' });
 
 export interface MaybeWithAuthorization {
-  authorization?: any;
+  authorization?: JWTPayload;
 }
 
-export interface AuthorizationContext extends Context, MaybeWithAuthorization {
+export interface AuthorizationContext extends ApiContext, MaybeWithAuthorization {
 }
 
-export type CheckToken = (authorization: any, ctx: Context) => Promise<boolean>;
+export type CheckToken = (authorization: JWTPayload | undefined, ctx: Context) => Promise<boolean>;
 
-export interface VerifyOptions {
-  complete: false;
-  ignoreExp?: boolean;
-  ignoreNbf?: boolean;
-  ignoreIat?: boolean;
-  maxTokenAge?: string;
-  subject?: string;
-  issuer?: string;
-  maxAuthAge?: string;
-  jti?: string;
-  clockTolerance?: string;
-  audience?: string | string[];
-  algorithms?: string[];
-  nonce?: string;
-  now?: Date;
-  crit?: string[];
+type KeySet = GetKeyFunction<JWSHeaderParameters, FlattenedJWSInput>;
+
+export interface Options {
+  verifyOptions?: JWTVerifyOptions;
   continueOnUnauthorized?: boolean;
 
   check?: CheckToken;
-  cacheOptions?: LRU.Options<string, Keys>;
-}
-
-export interface JwtHeader {
-  alg: string;
-  typ: 'JWT';
-  kid?: string;
-}
-
-export interface JwtPayload {
-  sub: string;
-  aub: string;
-  iss: string;
+  cacheOptions?: LRU.Options<string, KeySet>;
 }
 
 interface OpenidConfiguration {
-  jwks_uri: string;
-}
-
-interface Key {
-  /** The unique identifier for the key. */
-  kid: string;
-  /** The matching public key */
-  publicKey: string;
-}
-
-interface Keys {
-  [kid: string]: Key;
+  'jwks_uri': string;
 }
 
 export interface AuthorizationSuccess {
-  authorization: string | object;
+  authorization: JWTPayload;
   status: undefined;
 }
 
@@ -80,49 +50,75 @@ export interface AuthorizationFailure {
 }
 
 export namespace Jwt {
-  export function getBearerToken(headers: any): string | undefined {
+  export const getBearerToken = (headers: RequestHeaders): string | undefined => {
     const authorization = headers.authorization;
-    const matches = TOKEN_EXTRACTOR.exec(authorization);
+    if (!authorization) return undefined;
+    const matches = TOKEN_EXTRACTOR.exec(Array.isArray(authorization) ? authorization[0] : authorization);
     return matches && matches[1] || undefined;
-  }
+  };
+
+  export const decode = (jwt: string): {
+    payload: JWTPayload;
+    header: JWSHeaderParameters;
+  } => {
+    if (typeof jwt !== 'string') {
+      throw new Error('JWT must be a string');
+    }
+    const { 0: encodedHeader, 1: encodedPayload, length } = jwt.split('.');
+    if (length !== 3) {
+      throw new Error('Invalid JWT');
+    }
+    try {
+      const header = JSON.parse(Buffer.from(encodedHeader, 'base64').toString('utf8')) as JWSHeaderParameters;
+      const payload = JSON.parse(Buffer.from(encodedPayload, 'base64').toString('utf8')) as JWTPayload;
+      return { payload, header };
+    }
+    catch (err) {
+      throw new Error('Invalid JWT - Error Parsing JSON');
+    }
+  };
 }
 
 export class JwtDecoder {
-  public static getAuthorization(headers: any): any {
-    try {
-      // decode token
-      const token = Jwt.getBearerToken(headers);
-      if (token) {
-        return JWT.decode(token) || undefined;
-      }
-    } catch (err) {
-      debug.error(err);
-    }
-    return undefined;
-  }
-
   constructor() { }
 
-  public middleware(): Middleware & Router.IMiddleware {
-    return async (ctx: any, next: Application.Next): Promise<void> => {
-      ctx.authorization = JwtDecoder.getAuthorization(ctx.request.headers);
+  public middleware(): Middleware & Router.Middleware {
+    return async (ctx: AuthorizationContext, next: Application.Next): Promise<void> => {
+      const { headers } = ctx.request;
+      ctx.authorization = JwtDecoder.getAuthorization(headers);
       await next();
     };
   }
 }
 
-abstract class BaseJwtAuthentication {
-  constructor(
-    protected readonly verifyOptions?: VerifyOptions,
-  ) { }
-
-  public async getAuthorization(headers: any, logger: Logger): Promise<AuthorizationSuccess | AuthorizationFailure> {
+export namespace JwtDecoder {
+  export const getAuthorization = (headers: RequestHeaders): JWTPayload | undefined => {
     try {
       // decode token
       const token = Jwt.getBearerToken(headers);
       if (token) {
+        return Jwt.decode(token).payload;
+      }
+    } catch (err) {
+      debug.error(err);
+    }
+    return undefined;
+  };
+}
+
+abstract class BaseJwtAuthentication {
+  constructor(
+    protected readonly options?: Options,
+  ) { }
+
+  public async getAuthorization(headers: RequestHeaders, log: Logger): Promise<AuthorizationSuccess | AuthorizationFailure> {
+    try {
+      // decode token
+      const token = Jwt.getBearerToken(headers);
+      if (token) {
+        const result = await jwtVerify(token, await this.getSecretOrPublicKey(token), this.options?.verifyOptions);
         return {
-          authorization: JWT.verify(token, await this.getSecretOrPublicKey(token), this.verifyOptions),
+          authorization: result.payload,
           status: undefined,
         };
       }
@@ -131,21 +127,24 @@ abstract class BaseJwtAuthentication {
         status: Statuses.UNAUTHORIZED,
       };
     } catch (err) {
+      if (!(err instanceof Error)) throw err;
       switch (err.name) {
         case 'JWTClaimInvalid':
-          logger(err.message);
+        case 'JWSInvalid':
+        case 'JWSSignatureVerificationFailed':
+          log.warn(err.message);
           return {
             authorization: undefined,
             status: Statuses.FORBIDDEN,
           };
         case 'JWTMalformed':
-          logger(err.message);
+          log.warn(err.message);
           return {
             authorization: undefined,
             status: Statuses.UNAUTHORIZED,
           };
         default:
-          logger.error(err);
+          log.error(err);
           return {
             authorization: undefined,
             status: Statuses.UNAUTHORIZED,
@@ -154,14 +153,14 @@ abstract class BaseJwtAuthentication {
     }
   }
 
-  public middleware(): Middleware & Router.IMiddleware {
-    return async (ctx: any, next: Application.Next): Promise<void> => {
+  public middleware(): Middleware & Router.Middleware {
+    return async (ctx: AuthorizationContext, next: Application.Next): Promise<void> => {
       const { status, authorization } = await this.getAuthorization(ctx.request.headers, ctx.logger || debug);
       if (status) {
         ctx.status = status;
-        if (!this.verifyOptions?.continueOnUnauthorized) return;
+        if (!this.options?.continueOnUnauthorized) return;
       }
-      if (this.verifyOptions?.check && !(await this.verifyOptions.check(authorization, ctx))) {
+      if (this.options?.check && !(await this.options.check(authorization, ctx))) {
         ctx.status = Statuses.FORBIDDEN;
         return;
       }
@@ -170,62 +169,52 @@ abstract class BaseJwtAuthentication {
     };
   }
 
-  protected abstract getSecretOrPublicKey(token: string): Promise<ConsumeKeyInput>;
+  protected abstract getSecretOrPublicKey(token: string): Promise<KeyLike | JWTVerifyGetKey>;
 }
 
 export class JwtAuthentication extends BaseJwtAuthentication {
   constructor(
     private readonly secretOrPublicKey: string | Buffer,
-    verifyOptions?: VerifyOptions,
+    options?: Options,
   ) {
-    super(verifyOptions);
+    super(options);
   }
 
-  protected async getSecretOrPublicKey(): Promise<ConsumeKeyInput> {
-    return this.secretOrPublicKey;
+  protected getSecretOrPublicKey(): Promise<KeyLike | JWTVerifyGetKey> {
+    return Promise.resolve(createSecretKey(Buffer.from(this.secretOrPublicKey)));
   }
 }
 
 export class JwtJwksAuthentication extends BaseJwtAuthentication {
-  private cache: LRU<string, Keys>;
+  private cache: LRU<string, KeySet>;
 
   constructor(
-    verifyOptions?: VerifyOptions,
+    options?: Options,
   ) {
-    super(verifyOptions);
-    this.cache = new LRU<string, Keys>({
+    super(options);
+    this.cache = new LRU<string, KeySet>({
       max: 500,
       maxAge: Time.hours(1),
-      ...verifyOptions?.cacheOptions,
+      ...options?.cacheOptions,
     });
   }
 
-  protected async getSecretOrPublicKey(token: string): Promise<ConsumeKeyInput> {
-    const { header, payload } = JWT.decode(token, { complete: true }) as { header: JwtHeader, payload: JwtPayload };
-    if (!payload.iss.startsWith('https://') || !header.kid) throw new Error("Token doesn't support JWKS");
+  protected async getSecretOrPublicKey(token: string): Promise<KeyLike | JWTVerifyGetKey> {
+    const { header, payload } = Jwt.decode(token);
+    if (!payload.iss?.startsWith('https://') || !header.kid) throw new Error("Token doesn't support JWKS");
 
-    const keys = await this.getKeys(payload.iss);
-
-    const matchedKey = keys.hasOwnProperty(header.kid) ? keys[header.kid] : null;
-    if (!matchedKey) throw new Error(`Unable to ind matching key for iss=${payload.iss} kid=${header.kid}`);
-    return matchedKey.publicKey;
+    return this.getKeySet(payload.iss);
   }
 
-  private async getKeys(issuer: string): Promise<Keys> {
-    if (this.cache.has(issuer)) return this.cache.get(issuer)!;
+  private async getKeySet(issuer: string): Promise<KeySet> {
+    const cached = this.cache.get(issuer);
+    if (cached) return cached;
 
     const configUrl = new URL('./.well-known/openid-configuration', issuer).toString();
     const { data: configuration } = await axios.get<OpenidConfiguration>(configUrl);
 
-    const { data: jwks } = await axios.get<JSONWebKeySet>(configuration.jwks_uri);
-    const keyStore = JWKS.asKeyStore(jwks);
-    const keys = keyStore.all({ kty: 'RSA' })
-      .filter((key) => key.use === 'sig' && key.kid)
-      .map((key): Key => {
-        return { kid: key.kid, publicKey: key.toPEM() };
-      })
-      .reduce((acc, key) => ({ ...acc, [key.kid]: key }), {} as Keys);
-    this.cache.set(issuer, keys);
-    return keys;
+    const jwks = createRemoteJWKSet(new URL(configuration.jwks_uri));
+    this.cache.set(issuer, jwks);
+    return jwks;
   }
 }
